% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{postgres_query}
\alias{postgres_query}
\title{Query a Postgres database}
\usage{
postgres_query(con, query)
}
\arguments{
\item{con}{Connection to a PostgreSQL database.}

\item{query}{Character string, to be parsed as database query}
}
\description{
Query a Postgres database
}
\section{Prepared statements}{

The PostgreSQL backend supports the following prepared statement syntax:
\code{"$1"}
}

