% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{sqlite_query}
\alias{sqlite_query}
\title{Query a SQLite database}
\usage{
sqlite_query(con, query)
}
\arguments{
\item{con}{Connection to a SQLite database.}

\item{query}{Character string, to be parsed as database query}
}
\description{
Query a SQLite database
}
\section{Prepared statements}{

The SQLite backend supports the following prepared statement syntax:
\code{?}, \code{$1}, \code{:name}, \code{$name}
}

